Thinned out existing implementation to fit exact requirements, added tests, fixed structure.
