+++ src/main/java/com/group16b/ApplicationLayer/DTOs/DiscountPolicyDTO.java
@@ -1,6 +1,6 @@
 package com.group16b.ApplicationLayer.DTOs;
 
-import com.group16b.DomainLayer.Policies.DiscountPolicy;
+import com.group16b.DomainLayer.Policies.DiscountPolicy.DiscountPolicy;
 
 public class DiscountPolicyDTO {
