--- src/main/java/com/group16b/DomainLayer/Policies/DiscountPolicy.java
@@ -1,8 +1,0 @@
-package com.group16b.DomainLayer.Policies;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public interface DiscountPolicy {
-    double calculateDiscount(double originalPrice);
-}
