+++ src/main/java/com/group16b/DomainLayer/Policies/DiscountPolicy/DiscountPolicy.java
@@ -1,0 +1,5 @@
+package com.group16b.DomainLayer.Policies.DiscountPolicy;
+
+public interface DiscountPolicy {
+    double calculateDiscount(double originalPrice);
+}
