+++ src/main/java/com/group16b/DomainLayer/Policies/DiscountPolicy/SumDiscount.java
@@ -1,13 +1,13 @@
-package com.group16b.DomainLayer.Policies;
+package com.group16b.DomainLayer.Policies.DiscountPolicy;
 
 import java.util.List;
 
-public class MaxDiscount implements DiscountPolicy {
+public class SumDiscount implements DiscountPolicy {
     private List<DiscountPolicy> policies;
 
-    public MaxDiscount(List<DiscountPolicy> policies) {
+    public SumDiscount(List<DiscountPolicy> policies) {
         if (policies == null || policies.isEmpty()) {
-            throw new IllegalArgumentException("MaxDiscount must have at least one policy.");
+            throw new IllegalArgumentException("SumDiscount must have at least one policy.");
         }
         this.policies = policies;
     }
